UPDATE NumberRepository.java (REMOVE overlap)
