+++ src/main/java/baseball/Computer.java
@@ -1,8 +1,5 @@
 package baseball;
 
-import java.util.HashSet;
-import java.util.Set;
-
 /**
  * @ClassName Computer
  * 
@@ -15,21 +12,14 @@ import java.util.Set;
  * @EMAIL asqwklop12@gmail.com
  *
  */
-public class Computer implements NumberRepository {
+public class Computer extends NumberRepository {
 
 	private static final int NUMBER_RANGE = 9;
 
+	@Override
 	public int insert() {
 		return (int) (Math.random() * NUMBER_RANGE);
 	}
 
-	@Override
-	public Set<Integer> save() {
-		Set<Integer> set = new HashSet<>();
-		while (set.size() < REPOSITORYSIZE) {
-			set.add(insert());
-		}
-		return set;
-	}
 
 }
